Transaction functions

+++ src/main/java/model/dto/OrderDetail.java
@@ -11,5 +11,5 @@ public class OrderDetail {
     private String orderID;
     private String itemCode;
     private Integer orderQty;
-    private Integer discount;
+    private double discount;
 }
